requirements.txt
flask
flask_mysqldb
mysql-connector-python
requests
pyopenssl
